fix(mobile): show usage bars for system-default login
The mobile home screen and accounts page only rendered rate-limit usage
when a host had at least one Orca-managed Claude/Codex account
(`accounts.length > 0`), ignoring `snapshot.rateLimits` entirely. Users
who authenticate via the system-default `claude login` / `codex login`
path have zero managed accounts, so `accounts.list` returned a valid
snapshot with empty account arrays and live system-default rate-limit
data — but the UI filtered the host out and showed no usage section at
all.

Rate limits are fetched for the active target (which defaults to the
system-default `host` login) regardless of managed accounts, so the data
was present but hidden.

- Extract pure types/selectors into account-usage-state.ts (no RN
  imports) so the visibility logic is unit-testable; AccountUsage.tsx
  re-exports them to keep import sites stable.
- Add hasActiveProviderUsage()/hasRenderableUsage(): a provider is shown
  when it has a managed account OR a successful rate-limit fetch with
  window data for the active target. Providers with no credentials
  (unavailable/error, no windows) stay hidden — no spurious cards.
- Home screen: include such hosts in accountsHosts and render a
  "System default" row instead of returning null.
- Accounts page: show usage bars on the active "System default" row.
- Extract the accounts screen StyleSheet into accounts-screen-styles.ts
  to keep the file under the max-lines limit.
