+++ mobile/app/h/[hostId]/accounts.tsx
@@ -2,7 +2,6 @@ import { useEffect, useState, useCallback } from 'react'
 import {
   View,
   Text,
-  StyleSheet,
   Pressable,
   ScrollView,
   ActivityIndicator,
@@ -15,13 +14,15 @@ import { ChevronLeft, Check, RefreshCw, User } from 'lucide-react-native'
 import { loadHosts } from '../../../src/transport/host-store'
 import { useHostClient } from '../../../src/transport/client-context'
 import type { RpcSuccess } from '../../../src/transport/types'
-import { colors, spacing, typography, radii } from '../../../src/theme/mobile-theme'
+import { colors, spacing } from '../../../src/theme/mobile-theme'
+import { styles } from './accounts-screen-styles'
 import { ClaudeIcon, OpenAIIcon } from '../../../src/components/AgentIcons'
 import {
   type AccountsSnapshot,
   type ProviderKey,
   getActiveProviderRateLimits,
   getInactiveProviderUsage,
+  hasActiveProviderUsage,
   UsageBar
 } from '../../../src/components/AccountUsage'
 
@@ -149,6 +150,25 @@ export default function AccountsScreen() {
             <View style={styles.rowMain}>
               <Text style={styles.rowTitle}>System default</Text>
               <Text style={styles.rowSubtitle}>Use the agent's own login</Text>
+              {/* Why: when system default is the active selection, activeUsage
+                  holds the system-default login's rate limits — surface them
+                  here so non-managed users still see their usage. */}
+              {state.activeAccountId === null && hasActiveProviderUsage(activeUsage) ? (
+                <View style={styles.usageRow}>
+                  <UsageBar
+                    label="5h"
+                    usedPercent={activeUsage?.session?.usedPercent ?? null}
+                    unavailable={!activeUsage?.session && activeUsage?.status !== 'fetching'}
+                    loading={activeUsage?.status === 'fetching' && !activeUsage?.session}
+                  />
+                  <UsageBar
+                    label="7d"
+                    usedPercent={activeUsage?.weekly?.usedPercent ?? null}
+                    unavailable={!activeUsage?.weekly && activeUsage?.status !== 'fetching'}
+                    loading={activeUsage?.status === 'fetching' && !activeUsage?.weekly}
+                  />
+                </View>
+              ) : null}
             </View>
             <View style={styles.rowTrailing}>
               {state.activeAccountId === null ? (
@@ -285,138 +305,3 @@ export default function AccountsScreen() {
     </SafeAreaView>
   )
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: colors.bgBase
-  },
-  topRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingHorizontal: spacing.md,
-    paddingTop: spacing.sm,
-    paddingBottom: spacing.sm,
-    gap: spacing.sm
-  },
-  backButton: {
-    width: 36,
-    height: 36,
-    borderRadius: 18,
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  iconButton: {
-    width: 36,
-    height: 36,
-    borderRadius: 18,
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  titleWrap: {
-    flex: 1
-  },
-  heading: {
-    fontSize: 20,
-    fontWeight: '700',
-    color: colors.textPrimary
-  },
-  subheading: {
-    fontSize: typography.metaSize,
-    color: colors.textSecondary,
-    marginTop: 1
-  },
-  scroll: {
-    paddingHorizontal: spacing.lg,
-    paddingTop: spacing.sm
-  },
-  section: {
-    marginBottom: spacing.xl
-  },
-  sectionHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: spacing.sm,
-    marginBottom: spacing.sm
-  },
-  sectionHeading: {
-    fontSize: typography.metaSize,
-    fontWeight: '600',
-    color: colors.textSecondary,
-    textTransform: 'uppercase',
-    letterSpacing: 0.5
-  },
-  card: {
-    backgroundColor: colors.bgPanel,
-    borderRadius: radii.card,
-    overflow: 'hidden'
-  },
-  row: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingVertical: spacing.md,
-    paddingHorizontal: spacing.md + 2
-  },
-  rowPressed: {
-    backgroundColor: colors.bgRaised
-  },
-  rowMain: {
-    flex: 1,
-    gap: 4
-  },
-  // Why: fixed-width trailing slot so the usage bars in `rowMain` keep the
-  // same width whether or not the row is currently selected (otherwise the
-  // checkmark on the active account squeezes the bars narrower than the
-  // inactive rows above/below it).
-  rowTrailing: {
-    width: 24,
-    alignItems: 'flex-end',
-    justifyContent: 'center',
-    marginLeft: spacing.sm
-  },
-  rowTitle: {
-    fontSize: typography.bodySize,
-    fontWeight: '500',
-    color: colors.textPrimary
-  },
-  rowSubtitle: {
-    fontSize: typography.metaSize,
-    color: colors.textSecondary
-  },
-  separator: {
-    height: StyleSheet.hairlineWidth,
-    backgroundColor: colors.borderSubtle,
-    marginHorizontal: spacing.md
-  },
-  usageRow: {
-    flexDirection: 'row',
-    gap: spacing.md,
-    marginTop: 4
-  },
-  errorText: {
-    fontSize: typography.metaSize,
-    color: colors.statusRed
-  },
-  placeholder: {
-    paddingVertical: spacing.xl * 2,
-    alignItems: 'center',
-    gap: spacing.sm
-  },
-  placeholderText: {
-    fontSize: typography.bodySize,
-    color: colors.textSecondary
-  },
-  footerHint: {
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    gap: spacing.sm,
-    paddingHorizontal: spacing.sm,
-    paddingTop: spacing.sm
-  },
-  footerHintText: {
-    flex: 1,
-    fontSize: typography.metaSize,
-    color: colors.textMuted,
-    lineHeight: 18
-  }
-})
